Reorganize AuthZone
